Elevate privileges only if not in virtualenv
Generally you want to use user privileges as much
as possible.  When running pip in a virtualenv, the
default behavior should be to use user privileges.
This mod sets a new attribute in the Pip ExternPackager
is_virtualenv, from the config_manager.  This value
is then used to flag whether to elevate the pip call.